remove hal nomenclature to rebrand as a bsp
