+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "tomu-hal"
+name = "tomu"
 version = "0.1.0"
 authors = ["Nurahmadie <nurahmadie@gmail.com>"]
 edition = "2018"
@@ -7,7 +7,7 @@ edition = "2018"
 [dependencies]
 cortex-m = "0.5.8"
 efm32 = { git = "https://github.com/jacobrosenthal/efm32hg309f64-pac", features = ["rt"], package="efm32hg309f64-pac" }
-tomu-hal-macros = { path = "macros", optional = true }
+tomu-macros = { path = "macros", optional = true }
 embedded-hal = { version = "0.2.2", features = ["unproven"] }
 cast = {version = "0.2.2",default-features=false }
 
@@ -16,10 +16,10 @@ cast = {version = "0.2.2",default-features=false }
 [dev-dependencies]
 cortex-m-rt = "*"
 panic-halt = "*"
-tomu-hal-macros = { path = "macros" }
+tomu-macros = { path = "macros" }
 
 [target.'cfg(not(target_os = "none"))'.dev-dependencies]
 compiletest_rs = "0.3.17"
 
 [features]
-toboot-custom-config = ["tomu-hal-macros"]
+toboot-custom-config = ["tomu-macros"]
